+++ common/src/main/java/net/spell_power/api/MagicSchool.java
@@ -2,9 +2,28 @@ package net.spell_power.api;
 
 import net.minecraft.util.Identifier;
 import net.spell_power.SpellPowerMod;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.function.Consumer;
 
 public enum MagicSchool {
-    ARCANE, FIRE, FROST, HEALING, LIGHTNING, SOUL;
+    ARCANE, FIRE, FROST, HEALING, LIGHTNING, SOUL,
+    PHYSICAL_MELEE(new Identifier("generic.attack_speed"), SpellDamageSource.Configurator.MELEE);
+
+    @Nullable private final Identifier externalAttributeId;
+    private final Consumer<SpellDamageSource> damageSourceConfigurator;
+    public boolean isExternalAttribute() {
+        return externalAttributeId != null;
+    }
+
+    MagicSchool() {
+        this(null, SpellDamageSource.Configurator.MAGIC);
+    }
+
+    MagicSchool(@Nullable Identifier externalAttributeId, Consumer<SpellDamageSource> damageSourceConfigurator) {
+        this.externalAttributeId = externalAttributeId;
+        this.damageSourceConfigurator = damageSourceConfigurator;
+    }
 
     public static MagicSchool fromAttributeId(Identifier id) {
         return valueOf(id.getPath().toUpperCase());
@@ -15,7 +34,14 @@ public enum MagicSchool {
     }
 
     public Identifier attributeId() {
+        if (externalAttributeId != null) {
+            return externalAttributeId;
+        }
         return new Identifier(SpellPowerMod.ID, spellName());
+    }
+
+    public Consumer<SpellDamageSource> damageSourceConfigurator() {
+        return damageSourceConfigurator;
     }
 
     public int color() {
@@ -37,6 +63,9 @@ public enum MagicSchool {
             }
             case SOUL -> {
                 return 0x2dd4da;
+            }
+            case PHYSICAL_MELEE -> {
+                return 0xb3b3b3;
             }
         }
         assert true;