+++ common/src/main/java/net/spell_power/api/enchantment/SpellPowerEnchanting.java
@@ -1,5 +1,9 @@
 package net.spell_power.api.enchantment;
 
+import com.google.common.collect.HashMultimap;
+import com.google.common.collect.Multimap;
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.enchantment.Enchantments;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.item.ArmorItem;
 import net.minecraft.item.Item;
@@ -8,12 +12,13 @@ import net.minecraft.util.registry.Registry;
 import net.spell_power.api.MagicSchool;
 import net.spell_power.api.attributes.SpellAttributes;
 
-import java.util.ArrayList;
-import java.util.EnumSet;
-import java.util.List;
+import java.util.*;
+import java.util.function.BiFunction;
 import java.util.function.Function;
 
 public class SpellPowerEnchanting {
+
+    // MARK: Armor
     private static final ArrayList<ArmorItem> armor = new ArrayList<>();
 
     public static void registerArmor(ArmorItem... items) {
@@ -51,6 +56,17 @@ public class SpellPowerEnchanting {
         return false;
     }
 
+    // Defines connection between enchantments and powered schools
+    public record AttributeBoost(Enchantment enchantment, BiFunction<Double, Integer, Double> amplifier) { }
+    private static final Multimap<MagicSchool, AttributeBoost> powerMap = HashMultimap.create();
+    public static void boostSchool(MagicSchool school, Enchantment enchantment, BiFunction<Double, Integer, Double> amplifier) {
+        powerMap.put(school, new AttributeBoost(enchantment, amplifier));
+    }
+
+    public static Collection<AttributeBoost> boostersFor(MagicSchool school) {
+        return powerMap.get(school);
+    }
+
     public static EnumSet<MagicSchool> relevantSchools(ItemStack stack, EquipmentSlot slot) {
         var schools = EnumSet.noneOf(MagicSchool.class);
         var attributes = stack.getAttributeModifiers(slot);
@@ -63,5 +79,15 @@ public class SpellPowerEnchanting {
             }
         }
         return schools;
+    }
+
+    static {
+        for(var entry: Enchantments_SpellPower.damageEnchants.entrySet()) {
+            var enchantment = entry.getValue();
+            for (var school: enchantment.poweredSchools()) {
+                boostSchool(school, enchantment, enchantment::amplify);
+            }
+        }
+        boostSchool(MagicSchool.PHYSICAL_MELEE, Enchantments.SHARPNESS, (value, level) -> value * (1 + ((0.05) * level)));
     }
 }